Pycord Updates ruined cogs. Fixed them.
