+++ src/brvns_bot.py
@@ -3,22 +3,23 @@ BRVNS Bot API
 """
 import logging
 
-from discord.ext.commands import Bot
+import discord.ext.commands as ext_commands
 
+logging.basicConfig(
+    format="| %(asctime)s | %(levelname)-8s | [%(filename)s:%(lineno)d] | %(message)s"
+)
 logger = logging.getLogger()
 logger.setLevel("INFO")
 
 
 # pylint: disable=too-many-ancestors
-class BrvnsBot(Bot):
+class BrvnsBot(ext_commands.Bot):
     """
     BRVNS Bot
     """
 
     def __init__(self, debug_guilds):
         super().__init__(self, debug_guilds=debug_guilds)
-        self.load_extension("src.cogs", recursive=True)
-        logger.info("Cogs added")
 
     async def on_ready(self):
         """
